Started separating literal structs into files with individual type_check functions
